DRY out temporary fixture by moving it to conftest
